autoexec 增加参数 rexec.fail.reservedir

+++ lib/RunNode.py
@@ -1051,10 +1051,11 @@ class RunNode:
 
                             self._saveOpOutput(op)
                     try:
-                        if tagent.agentOsType == 'windows':
-                            tagent.execCmd(self.username, 'rd /s /q "{}"'.format(remoteRoot))
-                        else:
-                            tagent.execCmd(self.username, "rm -rf {}".format(remoteRoot))
+                        if ret == 0 or self.context.fialRserveDir == 0:
+                            if tagent.agentOsType == 'windows':
+                                tagent.execCmd(self.username, 'rd /s /q "{}"'.format(remoteRoot))
+                            else:
+                                tagent.execCmd(self.username, "rm -rf {}".format(remoteRoot))
                     except Exception as ex:
                         self.writeNodeLog('WARN: Remote remove directory {} failed {}\n'.format(remoteRoot, ex))
             except Exception as ex:
@@ -1265,7 +1266,8 @@ class RunNode:
                             self.writeNodeLog("ERROR: Download output failed {}\n".format(ex))
                             ret = 2
                     try:
-                        ssh.exec_command("rm -rf {}".format(remoteRoot, remoteRoot))
+                        if ret == 0 or self.context.fialRserveDir == 0:
+                            ssh.exec_command("rm -rf {}".format(remoteRoot, remoteRoot))
                     except Exception as ex:
                         self.writeNodeLog("WARN: Remove remote directory {} failed {}\n".format(remoteRoot, ex))
